+++ roqoqo-quest/src/backend.rs
@@ -200,7 +200,7 @@ impl Backend {
         };
 
         let (number_used_qubits, register_lengths) =
-            get_number_used_qubits_and_registers(&circuit_vec);
+            get_number_used_qubits_and_registers(&circuit_vec)?;
 
         if number_used_qubits > self.number_qubits {
             return Err(RoqoqoBackendError::GenericError { msg: format!(" Insufficient qubits in backend. Available qubits:`{}`. Number of qubits used in circuit:`{}` ", self.number_qubits, number_used_qubits) });